changing the FIXPNT template interface to use unsigned int instead of size_t for non-type parameters

+++ include/universal/number/fixpnt/manipulators.hpp
@@ -9,7 +9,7 @@
 namespace sw { namespace universal {
 
 // Generate a type tag for general fixpnt
-template<size_t nbits, size_t rbits, bool arithmetic, typename bt>
+template<unsigned nbits, unsigned rbits, bool arithmetic, typename bt>
 std::string type_tag(const fixpnt<nbits, rbits, arithmetic, bt>& = {}) {
 	std::stringstream s;
 	s << "fixpnt<"
